matrix-alert-bot/requirements.txt
matrix-nio[e2e]
requests
python-dotenv
structlog